Remove merge string process that is now managed by WPAndroid
From discussions in Slack, we no longer need to run this script when
generating the Gutenberg Mobile releases. The WPAndroid release process
generates the latest strings.

> The code freeze script gets the strings from
> https://github.com/wordpress-mobile/gutenberg-mobile/blob/develop/bundle/android/strings.xml,
> using the version on the release tag on GH. As long as that file is up
> to date on the release tag, we are fine.

One note is that this process currently only happens on initial code
freeze. Beta fixes do not pull in additional string changes. We should
avoid adding or modifying strings in beta fixes. If it is required, we
should discuss them with the Platform team.