Add clarity regarding Aztec releases
There has been some confusion as to when Aztec releases need to be
created. Not every change in Aztec requires a new release during the
Gutenberg Mobile release process. Only changes that have an impact on
the Gutenberg project itself warrant a new Aztec release.

Ideally the need for a new Aztec release is identified by a change to
the Aztec version references found in the Gutenberg repository. If those
reference do not point to a stable, tagged release, then a new Aztec
release is likely required.