Tweak order of steps under "Create the Release"
The release notes need to be updated in both the Gutenberg and Gutenberg Mobile repos before the PRs are ready for review. As these updates require new commits, it makes sense to perform this step before waiting for tests to complete. This commit, therefore, suggests moving this step before the steps to run tests under "Create the Release".