Update "Changes" section in PR template
With this commit, the "Changes" section in the gutenberg-mobile PR template is updated. It now prompts the release wrangler to link to the GitHub issues related to a change, in addition to the PRs. The aim is to make the reason for each change clearer to those viewing the PR.