Update test plan to only include smoke testing
Previously we did extensive testing on the Gutenberg Mobile release before it is merged into the main apps. Nowadays, we only do minimal testing of release to ensure the editor is not broken, then make the release, then do more thorough testing of the editor that's included in the main app beta builds.

This updates the release PR templates to reflect this change.