Add step to run `pod install` in XCFramework folder to automation
This is necessary after the introduction of the XCFramework builder
project in
https://github.com/wordpress-mobile/gutenberg-mobile/pull/5739 and
https://github.com/wordpress-mobile/gutenberg-mobile/pull/5740.

Worth noting that the XCFramework distribution pipeline and integration
in Jetpack and WordPress iOS is currently incomplete. The step is
required to avoid a CI failure but has not effect on the quality of the
final release. This will hopefully change within the next couple of
releases.